refactor :: student 관련 package 구조 변경
ddd 에 맞춰서 각각 domain 패키지를 추가한 뒤 entity, repository를 이동시켰습니다.
