+++ src/main/java/com/keepgoing/keepserver/domain/student/domain/repository/StudentRepository.java
@@ -1,6 +1,6 @@
-package com.keepgoing.keepserver.domain.student.repository;
+package com.keepgoing.keepserver.domain.student.domain.repository;
 
-import com.keepgoing.keepserver.domain.student.entity.Student;
+import com.keepgoing.keepserver.domain.student.domain.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
