+++ src/main/java/com/keepgoing/keepserver/domain/student/domain/repository/dto/StudentRequestDto.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.student.repository.dto;
+package com.keepgoing.keepserver.domain.student.domain.repository.dto;
 
 import lombok.Getter;
 import lombok.Setter;
